fix: Refactor Task Queue for DP Expiration
