+++ src/main/java/it/gov/pagopa/pu/workflow/wf/debtposition/expirationdp/config/CheckDebtPositionExpirationWfConfig.java
@@ -4,8 +4,6 @@ import io.temporal.workflow.Workflow;
 import it.gov.pagopa.payhub.activities.activity.debtposition.DebtPositionExpirationActivity;
 import it.gov.pagopa.pu.workflow.config.temporal.BaseWfConfig;
 import it.gov.pagopa.pu.workflow.config.temporal.TemporalWFImplementationCustomizer;
-import it.gov.pagopa.pu.workflow.wf.debtposition.expirationdp.activity.ScheduleCheckDpExpirationActivity;
-import it.gov.pagopa.pu.workflow.wf.debtposition.expirationdp.wfexpiration.CheckDebtPositionExpirationWFImpl;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
 
@@ -18,9 +16,4 @@ public class CheckDebtPositionExpirationWfConfig extends BaseWfConfig {
       TemporalWFImplementationCustomizer.baseWfConfig2ActivityOptions(this));
   }
 
-  public ScheduleCheckDpExpirationActivity buildScheduleCheckDpExpirationActivityStub() {
-    return Workflow.newActivityStub(ScheduleCheckDpExpirationActivity.class, TemporalWFImplementationCustomizer.baseWfConfig2ActivityOptions(
-      CheckDebtPositionExpirationWFImpl.TASK_QUEUE_CHECK_DP_EXPIRATION_LOCAL_ACTIVITY,
-      this));
-  }
 }
